Disable world input when the mouse is over UI

+++ src/model/ui/mod.rs
@@ -1,6 +1,6 @@
 use nalgebra::{Point2, Vector2};
 
 pub struct Button {
-    pub position: Point2<f32>,
-    pub size: Vector2<f32>,
+    pub position: Point2<i32>,
+    pub size: Vector2<i32>,
 }
